backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
pandas
scikit-learn
numpy
python-dotenv